pyproject.toml
[tool.poetry]
name = "python-camera-classifier"
version = "0.1.0"
description = "Python Camera Classifier"
authors = ["Anderson F Lima <andersonfaustinolima@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^8.1.0"
sklearn = "^0.0"
numpy = "^1.19.4"
opencv-python = "^4.5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
